Added getImageUrl tests
