--- frontend/.babelrc
@@ -1,1 +1,0 @@
-{ "presets": ["@babel/preset-env"] }
