+++ frontend/package.json
@@ -25,7 +25,6 @@
     "vue": "^2.6.11",
     "vue-axios": "^2.1.5",
     "vue-clipboard2": "^0.3.1",
-    "vue-lazyload": "^1.3.3",
     "vue-moment": "^4.1.0",
     "vue-router": "^3.2.0",
     "vue2-gmap-custom-marker": "^5.6.0",
